move InternalLink to shared/components
