+++ src/containers/Accounts/AccountTransactionsTable/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { translate } from 'react-i18next';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import Link from '../../shared/InternalLink';
+import Link from '../../shared/components/InternalLink';
 import { ReactComponent as SuccessIcon } from '../../shared/images/success.svg';
 import { ReactComponent as FailIcon } from '../../shared/images/ic_fail.svg';
 import { localizeDate, concatTx } from '../../shared/utils';
