+++ src/containers/Network/NodesTable.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { translate } from 'react-i18next';
-import Link from '../shared/InternalLink';
+import Link from '../shared/components/InternalLink';
 import Loader from '../shared/components/Loader';
 import { durationToHuman } from '../shared/utils';
 import './css/nodesTable.css';
