+++ src/containers/Transactions/Simple/EscrowFinish.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Link from '../../shared/InternalLink';
+import Link from '../../shared/components/InternalLink';
 import Account from '../../shared/components/Account';
 import Currency from '../../shared/components/Currency';
 import { CURRENCY_OPTIONS } from '../../shared/transactionUtils';
