+++ src/containers/Validators/Simple/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Link from '../../shared/InternalLink';
+import Link from '../../shared/components/InternalLink';
 
 const Simple = props => {
   const { t, data /* , language */ } = props;
