+++ src/containers/shared/components/Account.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Link from '../InternalLink';
+import Link from './InternalLink';
 
 interface Props {
   link?: boolean;
